pyproject.toml
[project]
name = "datacampagent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "browser-use>=0.12.0",
    "langchain-openai>=1.1.9",
    "playwright>=1.58.0",
    "python-dotenv>=1.2.1",
]
